Refactored to not drop forms in StandardOrth transcription
The clean_form method contained logic to make sense of the forms
given the specified Transcription. This was in order to help with
subsequent segmentation. Given the huge amount of variation even
after this kinf of normalization, we give up on segmentation for
the time being, and instead deliver the forms and the corresponding
Transcription spec *as is* in the CLDF output.